modified

+++ src/BIT_WISE/BitWiseOperator.java
@@ -2,7 +2,7 @@ package BIT_WISE;
 
 public class BitWiseOperator {
     public static void main(String[] args) {
-       int x=5;//00000000000000000000000000110010  == 5
+       long x=5;//00000000000000000000000000110010  == 5
        x=~x;   //11111111111111111111111111001101  == -6
         System.out.println(x); // -6
   /*
